Cmpletely silence the default ::Rack::CommonLogger if there is a Rackstash::Rack::Middleware
Under some environments, Rack adds a default Rack::CommonLogger
middleware to its default stack. Since this middleware logs the same
information we log with Rackstash but with a potentially unexpected
format, we completely silence the CommonLogger middleware so that it
doesn't log anything as long as there is a
Rackstash::Sinatra::Middleware in the current middleware stack.

If you want to still log in the CLF format of the CommonLogger
middleware, you can use the CLF encoder of Rackstash.