Explicitly set Sinatra::Base#logging in tests
Since the default value is environment-specific (it is derived from the
SINATRA_ENV or RACK_ENV environment variables), we need to explicitly
set it for repeatability of the tests.

The tests have failed on Travis CI, it sets RACK_ENV=test. This disables
Sinatra logging for Sinatra::Application classes by default.